overo: Support Arbor50C/70C expansion boards
Arbor50C and Arbor70C are versions of the Arbor43C expansion board for
the Gumstix Overo COM that support larger touchscreens (5" and 7"
respectively).  As a few peripherals are wired differently, Linux
uses a different device tree for each.  Detect the onboard EEPROM to
detect the board and pass the correct devicetree name to Linux.

Signed-off-by: Ash Charles <ashcharles@gmail.com>